pyproject.toml
[project]
name = "vibecurator-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets==4.5.0",
    "openai-agents==0.6.9",
    "pip==25.3",
    "python-dotenv>=1.0.0",
    "qdrant-client==1.16.2",
    "sentence-transformers==5.2.0",
    "streamlit>=1.39.0",
    "tqdm>=4.66.3",
]
